feat(uniq): change int to uint8 for padding len, change appropriate tests

+++ stringutil/pad.go
@@ -3,18 +3,18 @@ package stringutil
 import "strings"
 
 // PaddingLeft add padWith to the left of s until it reaches a length overallLen
-func PaddingLeft(s string, padWith string, overallLen int) string {
-	if overallLen < len(s) {
+func PaddingLeft(s string, padWith string, overallLen uint8) string {
+	if int(overallLen) < len(s) {
 		return s
 	}
 
 	res := paddingString(s, padWith, overallLen) + s
-	return res[(len(res) - overallLen):]
+	return res[(len(res) - int(overallLen)):]
 }
 
 // PaddingRight add padWith to the right of s until it reaches a length overallLen
-func PaddingRight(s string, padWith string, overallLen int) string {
-	if overallLen < len(s) {
+func PaddingRight(s string, padWith string, overallLen uint8) string {
+	if int(overallLen) < len(s) {
 		return s
 	}
 
@@ -23,8 +23,8 @@ func PaddingRight(s string, padWith string, overallLen int) string {
 	return res[:overallLen]
 }
 
-func paddingString(s string, padWith string, overallLen int) string {
-	padCount := 1 + (overallLen-len(s))/len(padWith)
+func paddingString(s string, padWith string, overallLen uint8) string {
+	padCount := 1 + (int(overallLen)-len(s))/len(padWith)
 	if padCount < 0 {
 		padCount = 0
 	}
